requirements.txt
asgiref==3.2.10
certifi==2021.5.30
charset-normalizer==2.0.1
DBUtils==2.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.1.7
django-crispy-forms==1.12.0
django-formtools==2.2
django-import-export==2.5.0
django-redis==5.0.0
django-reversion==4.0.0
djangorestframework==3.9.1
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
httplib2==0.9.2
idna==3.2
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
reversion==0.2
ronglian-sms-sdk==1.0.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
xadmin @ https://codeload.github.com/sshwsfc/xadmin/zip/django2
xlrd==2.0.1
xlwt==1.3.0
